SimpleLottery: Renaming setup.rs params.rs
