+++ src/simple_lottery/mod.rs
@@ -3,7 +3,7 @@
 
 pub mod init;
 
-pub mod setup;
+pub mod params;
 
 pub mod proof;
